Throw an error message for invalid tickers + allow decimal values for sum to invest (#6)
* Throw an error message for invalid tickers
- Print an error message when the portfolio CSV-file contains tickers that
don't exist on Yahoo finance.
- Allow decimal values to be used for the sum to invest input.

* Bump up version number
Bump up version number in anticipation for version 1.0.1.

+++ portfolio_rebalancer/cli.py
@@ -13,7 +13,7 @@ def portfolio_rebalancer():
                                     "your portfolio as close as possible to "
                                     "your target allocation given a sum to "
                                     "invest."))
-@click.argument('sum-to-invest', type=click.INT)
+@click.argument('sum-to-invest', type=click.FLOAT)
 @click.option('-p', '--portfolio', 'portfolio_csv', required=True,
               help="CSV-file containing the portfolio and target allocations.")
 def calc(sum_to_invest, portfolio_csv):
